Peel "refs/heads" for revision parsing.
Gitiles does not support having a branch name as "refs".
Fix the parser to handle it correctly.

When asking for "refs/heads/master", Gitiles will try to find out which part of this string is a branch and which part is a a file path (as appending a file path directly to a revision/branch seems to be allowed in Gitiles). To do so, it tests "refs", "refs/heads", and "refs/heads/master" exactly in that order (in RevisionParser#parse). If one of the strings turns out to be a valid branch, Gitiles assumes that the remaining string suffix is the file path. The repository platform/nlmanifest contains the branch "refs" (or in its longer form "refs/heads/refs"). Due to this, Gitiles will recognize "refs" as a valid branch and assume that "/heads/master" is a file path. As there is no file "heads/master" in the branch "refs", you get a 404.

PiperOrigin-RevId: 266266027
Change-Id: Ie874200c5fff2d3de64757b67e0b43a90dfb1690