Peel "refs/heads" for revision parsing.
Gitiles does not support having a branch name as "refs".
Fix the parser to handle it correctly.

If you create a branch name called "refs" and in try to view that in Gitiles like below:
https://host.googlesource.com/repo/+/refs/heads/refs
Gitiles will recognize "refs" as a valid branch and assume that "/heads/master" is a file path.
As there is no file "heads/master" in the branch "refs", you get a 404.
instead of repo=refs/heads/refs, path = ""

The tests in this review creates a new branch called "refs/heads/refs" and then parses the same to verify the parser.

PiperOrigin-RevId: 266266027
Change-Id: Ie874200c5fff2d3de64757b67e0b43a90dfb1690