Update patch set 5
Patch Set 5:

(4 comments)

Sorry for dropping the ball on this one. Refreshed and gitiles is now using JGit 5.4, so should be ready to go in (if we are happy with everybody getting topoSort reachability checks).

In a follow up I'll update it to JGit 5.6 (that receives Stream instead of Collection).

Patch-set: 5
Reviewer: Gerrit User 1072453 <1072453@173816e5-2b9a-37c3-8a2e-48639d4f1153>