Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > > > This conflicts with https://gerrit-review.googlesource.com/c/gitiles/+/231034
> > > 
> > > I believe they are both valid checks and they both are pointing at avoiding to render too much content.
> > > Can we keep both?
> > 
> > I believe David is saying we'd like help with the merge to "master".
> > 
> 
> Yes, but I was also asking because I wasn't sure if they were both fixing the same thing in different ways.

I believe yes, they are both trying to prevent the JVM to explode when trying to render big files.
The fix on master is counting the number of lines, which isn't necessarily solving the underlying problem, as you may have an XML without new lines and you'll still blow up the heap.

> 
> > That reminds me --- I should have asked, why is this targeted at the stable-0.2 branch instead of "master"?
> 
> It was put on stable-0.2 with the intention of using it with gerrit 2.16 which is currently using 0.2-7.  See the discussion on https://gerrit-review.googlesource.com/c/plugins/gitiles/+/246077

Yes, we still need to find a way to apply the fix to Gerrit :-(

Patch-set: 7