Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > > Bazel tests failing locally, also before this change.
> > 
> > Since v0.3-3 that is. Before this tag, I cannot build with bazel locally, likely because using bazel 2.0.0. Are bazel tests a VR+1 criteria though for gitiles? Maybe not as much as passing them in Eclipse; I haven't tried that yet with this change.
> 
> My local gitiles server works with basic or limited browsing in it; looks fine. However running the tests in eclipse fails for almost half of them, similar to bazel's.

The tests failing in Eclipse is a known issue.  See issue 10843 for a workaround.

Tests are passing for me with bazel on the command line with the latest PS which is based on the bazel version and bazlets revision updates. I haven't tried it before doing those updates.

Patch-set: 3
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>