LogServlet: use RevSort.TOPO_KEEP_BRANCH_TOGETHER
For the "topo-order" parameter in +log, use
RevSort.TOPO_KEEP_BRANCH_TOGETHER instead of RevSort.TOPO. Using the
TOPO_KEEP_BRANCH_TOGETHER flag causes JGit RevWalk to return commits
in topological order without intermixing lines of commit history. This
is an alternative to using RevSort.TOPO, which does not guarantee that
lines of history are not intermixed.

Signed-off-by: Alex Spradlin <alexaspradlin@google.com>
Change-Id: Ic59bfb8e1526615b261c0fac0b57aad107e3fa7d