Update patch set 1
Patch Set 1:

Sorry for the slow reply. Before using this on chromium.googlesource.com, I'd like to make sure that it is also supported at source.chromium.org. Otherwise, people will have a bad experience when switching between the two tools.

That's not in itself a reason not to merge this change --- we could have this support in Gitiles and just have it disabled on that host --- but it's why I've been slow to review.

Patch-set: 1
Reviewer: Gerrit User 1011842 <1011842@173816e5-2b9a-37c3-8a2e-48639d4f1153>