Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Sorry for the slow reply. Before using this on chromium.googlesource.com, I'd like to make sure that it is also supported at source.chromium.org. Otherwise, people will have a bad experience when switching between the two tools.
> 
> That's not in itself a reason not to merge this change --- we could have this support in Gitiles and just have it disabled on that host --- but it's why I've been slow to review.

We're working on a large documentation project for Chromium UI and plan to point folks straight to gitiles for information as Chromium already has a large source of documentation there. 

source.chromium.org is still not the primary place for documentation and so is lower on our priority list.

As a result, we'd really like this support and if it turns out to be useful, we can push source.chromium.org to have it.

Patch-set: 1
CC: Gerrit User 1073824 <1073824@173816e5-2b9a-37c3-8a2e-48639d4f1153>