Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > As a result, we'd really like this support and if it turns out to be useful, we can push source.chromium.org to have it.
> 
> Hm, that doesn't line up with how ChOps folks have been handling this. Let's talk internally, since this aspect isn't about the Gitiles project.

I agree that we need to talk internally about this as to whether we want to land it.

It might be useful to review this change in parallel, if that's reasonably doable, just in case you want significant changes to this, so that Wei could work on said changes. I would not take such a review as indication that we'll in fact land and enable the feature.

If you would prefer to not do so until we have a path to turning this feature on on chromium.googlesource.com, though, I totally understand and that's fine by me. And, of course, Wei might also agree that there's no point in reviewing this until we're sure we want to land and enable it :)

Patch-set: 1
CC: Gerrit User 1019413 <1019413@173816e5-2b9a-37c3-8a2e-48639d4f1153>