Update patch set 1
Patch Set 1:

(1 comment)

I haven't reviewed it in detail but noticed that many of the changed files are not formatted consistently with the rest of the code. The project is using google-java-format; please run that over all the changes.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>