Drop u-pre weight from 500 to normal
When displaying file contents, as for examle on

https://gerrit.googlesource.com/gerrit/+/589af813ccca4c329bd4845852cd9d6d73d1d11d/gerrit-main/src/main/java/Main.java

u-pre (setting font weight to 500) is combined with u-monospace
(setting font-family to `Source Code Pro`).

The @import of `Source Code Pro` however only defines the font face for
font weight 400. So on default setups, the 500 falls back to using
`Source Code Pro` with the lighter weight 400 (i.e.: normal).

On setups, where the monospace font is provided with both 400 and 500
weights (e.g.: users having the fonts installed locally, or using
Gerrit's own provided fonts setup), the heavier 500 weight got used,
and lead to a unwarranted bold look of the source code.

By setting u-pre to normal, we do not change the look on default setups
yet allow to avoid bold-looking source code on setups that provide
their own fonts.

Change-Id: I7322c2762f6bce1a6fc6b0de1a2a3674092ef8a1