Set correct target url for symlinks
When attemting to resolv targetUrl for a symlink dirname was calculated
incorrectly.
After removing all trailing slashes the index of the last '/' was
calculated in an effort to get the dirname part from what was before
this last slash.
When attempting to remove the last slash and everything after it the
index was unnecesarily decremented by one which lead to a faulty
targetUrl and a faulty link to the symlink target in the UI.

   path = "foo/bar"
   lastSlashIndex = 3
   dirPath = "foo/bar".substring(0, 3-1) => "fo"

So if symlink "./foo/bar" had target "./foo/bar.md" the link to target
would be "./fo/bar.md".

Gitiles Bug: https://github.com/google/gitiles/issues/144

Change-Id: I14d6631b2ba7879ca66c091b90daf18e8763f314