Support branch redirect in Gitiles
Github has updated the default branch from master to main.
To have a smooth transition for users having link to an old branch, can now be redirected to a new branch.

BranchRedirectFilter primarily does the work and provides an override-able method getRedirectBranch which can be implemented to provide the branch redirect if available for the branch.

PiperOrigin-RevId: 348239020
Change-Id: I80729c8f315ca80f874e07d6919734cb07120625