Revert "Upgrade soy to 2020-08-24"
This reverts commit 742c739f96ae0546bd1e213a2603332dcebf42e5.

Han-Wen reported this soy version to show performance issues in Gerrit:

"Google is tracking this internally as b/176094888; hopefully we can get
it fixed in Soy proper.

The problem is that template rendering went from 9ms to 30ms. We render
quite a few templates for a single email, so adds ~120ms to each
operation that sends out an email.

It looks like this caused a significant slowdown on the benchmark under
contrib/benchmark-createchange.go (using reftable as refstorage)."

Hence reverting this update until it's fixed upstream.

See https://gerrit-review.googlesource.com/c/gerrit/+/280196

Change-Id: I17719d8dcbe08e63bd1c96b2fa011d9c643e926c