Parse large content in batches
BlobSoyData uses java-prettify library for content highlighting.
As a part of the content tokenizing java-prettify library uses
a standard java regular expression package java.util.regex. Java
regex engine is using a backtracking implementation strategy:
when a pattern presents two alternatives such as a|b, the engine
will try to match subpattern a first, and if that yields no match,
it will reset the input stream and try to match b instead. If
such choices are deeply nested, this strategy requires an
exponential number of passes over the input data before it can
detect whether the input matches. For certain input content the
java.util.regex matcher may run forever, or exceed the available
stack space and fail. Another issue is that the java-prettify library
uses some regular expressions that cannot be implemented in linear
time, for example before text matching `?=` which is using a
backtracking implementation strategy. This means that we are not
able to replace java regex engine with a regular expression engine
that runs in linear time, for example RE2/J. To mitigate the
StackOverflowException issue split content into multiple batches,
process each batch separately and combine the results.

Bug: Issue 14914
Change-Id: I25b0f1d8f5eeca6b9bf0e0a5048f5c2013963ebf