Link to JGit source as git submodule
Using a submodule to bring in the source files vastly simplifies
updating JGit. Gitiles developers can bump the SHA-1 of the submodule
and not deal with the Maven rebuild process for JGit.

This follows the approach taken in Gerrit in I16370e57f27.

To perform this migration, some build rules have to be renamed to meet
expectation in JGit bazel build.

Add JGit test dependencies to ensure that build of jgit tests  succeeds.

Moreover for stamping of JGit artifact, revision version must be exposed
in workspace status command. Use this opportunity to synchronize status
command between gerrit and gitiles repositories and migrate shell script
to python.

Change-Id: Ib35f7bad8b8f61e73f416937a387e3675bf1e3a2