Support branch redirecting from a deleted or non-existent branch to HEAD.
BranchRedirectFilter uses the ViewFilter to obtain the GitilesView. The GitilesView is not created for a deleted branch. This causes `BranchRedirectFilter to not redirect as expected. Since branch redirect is used primarily to migrate from master to main, and master to be eventually deleted. HEAD would be set to main since master is being deleted. This makes HEAD a correct place to redirect to.

PiperOrigin-RevId: 411156730
Change-Id: I4193a86d36e3d46cb19abf9c7b52a94b1feef5c1