Redo branch redirect
This version of BranchRedirect isn't a BranchRedirectFilter. Instead
this carries just redirects and is applied when a ViewFilter is created,
i.e. the very first step in the process.

The older implementation returned a redirect with a new URL. While this
does the redirect in-place. User does not know if the gitiles command
they requested is transformed to use a new branch. This transformation
is only valid on requests that are not from automation (i.e. HTTP only).

This will also help when the source branch is deleted. Older
implementation didn't work since ViewFilter wasn't created on deleted
or non-existent branch.

PiperOrigin-RevId: 411156730
Change-Id: I4193a86d36e3d46cb19abf9c7b52a94b1feef5c1