Bazel: Switch to using toolchain resolution for java rules
Bump Bazel version to release 5.0.0. In this new Bazel release,
--incompatible_use_toolchain_resolution_for_java_rules is flipped, that
means that the build must be adapted to toolchain resolution.

Given that new Bazel release also added support for remote JDK 17, add
support for building with remote JDK 17 to produce major byte code
version 61. Note that some tests are still failing on JDK 17. This will
be fixed in follow-up change.

Change-Id: I5bc5f640bcd43529f799a8df01ad4a496f0394d3