doc: build rules requires /usr/bin/python
maven_jar would fail to fetch material when a `python` executable is not
found in path. It looks it up with:

 python = ctx.which("python")

Since Debian 11 and Ubuntu 20.04 LTS, there is no more a
`/usr/bin/python` to get rid of the ambiguity between python 2 and
python 3.

8cc50228 erroneously mentionned the build rule requires python 2,
instead of the lack of `python`.  Adjust the document to instruct people
to install Python 3.

Change-Id: I8c56a79735cc87e6a2e2721c5a23feec8fb2dc3b