Bump Jetty version to 11.0.20
Eclipse Jetty 11 has been released and certified as compatible with the
Jakarta Servlet v5.0 specification, released as part of Jakarta EE 9:
[1]. This new version of the Servlet specification uses the new jakarta
namespace, instead of the old javax.servlet namespace.

To update Gerrit to Jetty 11, we have to replace servlet-api with
jakarta servlet API.

JGit was updated in this pending change: [2] and JGit submodule was
updated to point to it.

[1] https://jakarta.ee/release/9
[2] https://eclipse.gerrithub.io/c/eclipse-jgit/jgit/+/189213

Change-Id: Ia67f142a82f7f316220d18df7bf79e115d36e024