VisibilityChecker#isReachableFrom: Remove unused param and exceptions
The current signature of #isReachableFrom has couple of wrinkles: the
description is not used and the only caller of this method doesn't
differentiate the subclasses of IOException.

Fixing the method now, before deleting the deprecated variant, helps
downstream transition: subclasses can move to the updated API and
ignore the deprecated variant. If we want to change this later,
downstream needs to update at the same time than the superclass.

Update also the tests, to use the "new" variant.

Change-Id: If31043ab711a7a287ebf5529e2823fc8ed73ed97