Region: add committer data into the blame region
The blame region class contains author data about the blame candidate
associated to the region, but it does not contain the committer data.
Since committer can be different from the author, having committer data
is helpful with determining the person committed the change into
mainline.

Add a field for the committer in the Region class and populate it. The
BlameGenerator already supplies the data. For previously serialized
instances without the committer field, deserializing will result in a
null committer field; no action required.

Change-Id: I2370ad88306db83110bb25d408c176a12e036b7b
Signed-off-by: Xing Huang <xingkhuang@google.com>