Tests: update path prefix in git urls
When testing URLs, it is helpfull to distinguish between the base URL
(which can contains a path prefix such as `/r`) and the application
paths that is appended to it.

Using `/foo` is a bit meaningless, replace it by `/path-prefix` which is
self describing and helps differentiating between the base URL and the
rest of the path.

Change-Id: I91230750cc67b7a2876e4da7779a0b0e20f62639