Fix edit link when base URL has a path
On a setup having a path prefix such as `/r` the edit link woud be
broken because the prefix is stripped.

The reason is `URI.resolve()` is given an absolute path which strip the
path from the base URI:

  jshell> new URI("https://example.org/r").resolve("/admin/repos/edit");
                                      ^^
  $2 ==> https://example.org/admin/repos/edit
                           ^^^____oops

Which is consistent with python `os.path`:

  >>> os.path.join('https://example.org/r', '/admin/repos/edit');
  '/admin/repos/edit'

Remove the leading slash from the edit URI and update the test with the
proper expected value.

Bug: https://phabricator.wikimedia.org/T405481
Change-Id: I98e9d89cbc4850270f81b52665f2c8849f25c879